feat: load commands from dedicated folder
